requirements.txt
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-community==0.3.14
langchain-core==0.3.35
langgraph==0.2.66
python-dotenv==1.0.1
python_dateutil >= 2.5.3
setuptools >= 21.0.0
urllib3 >= 1.25.3, < 2.1.0
pydantic >= 2
typing-extensions >= 4.7.1
flask==3.1.0
flask-cors==5.0.0
google-generativeai==0.8.4
streamlit==1.42.0
streamlit-card
pillow